quotes_labels: Restrict quotes
This commit brings the restriction for the values that will not be allowed to have quotes
