+++ libnetdata/url/url.c
@@ -59,32 +59,6 @@ char url_percent_escape_decode(char *s) {
 }
 
 /**
- * Get byte length
- *
- * This (utf8 string related) should be moved in separate file in future
- *
- * @param c is the utf8 character
- *  *
- * @return It reurns the length of the specific character.
- */
-char url_utf8_get_byte_length(char c) {
-    if(!IS_UTF8_BYTE(c))
-        return 1;
-
-    char length = 0;
-    while(likely(c & 0x80)) {
-        length++;
-        c <<= 1;
-    }
-    //4 byte is max size for UTF-8 char
-    //10XX XXXX is not valid character -> check length == 1
-    if(length > 4 || length == 1)
-        return -1;
-
-    return length;
-}
-
-/**
  * Decode Multibyte UTF8
  *
  * Decode % encoded UTF-8 characters and copy them to *d
